# Small memory optimizations in badger write-path