## Which problem is this PR solving?

This is work from another unfinished PR ported as standalone. Adds benchmarks for profiling purposes and removes some unnecessary duplicate processing and memory pressure from the write-path mostly. 

Improves perf slightly:

```
benchmark                                           old ns/op      new ns/op      delta
BenchmarkWrites-8                                   8365954405     7515525425     -10.17%
BenchmarkServiceTagsRangeQueryLimitIndexFetch-8     61905754       59933936       -3.19%
BenchmarkServiceIndexLimitFetch-8                   41341527       39249518       -5.06%
```

But more interesting, reduces memory pressure towards GC when writing:

```
benchmark             old allocs     new allocs     delta
BenchmarkWrites-8     49572552       36456074       -26.46%

benchmark             old bytes      new bytes      delta
BenchmarkWrites-8     2080032376     1853283416     -10.90%
```

## Short description of the changes

time-package calls are reduced, bytes.Buffer is replaced with writing to slice directly and copy() is used instead of append( ...)